📝Modify jest configuration
