+++ jest.config.js
@@ -1,20 +1,15 @@
-// jest.config.js
-const { pathsToModuleNameMapper } = require('ts-jest/utils');
-// In the following statement, replace `./tsconfig` with the path to your `tsconfig` file
-// which contains the path mapping (ie the `compilerOptions.paths` option):
-const { compilerOptions } = require('./tsconfig');
 const config = {
   rootDir: '.',
   preset: 'ts-jest',
   roots: ['<rootDir>/__tests__'],
-  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.ts$',
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.(ts?)$',
   moduleFileExtensions: ['ts', 'js'],
   testEnvironment: 'node',
   moduleNameMapper: {
-    '^@components/(.*)$': '<rootDir>/src/components/$1',
+    '@components/(.*)$': '<rootDir>/src/components/$1',
   },
   collectCoverage: true,
-  collectCoverageFrom: ['**/*.{ts}', '!**/node_modules/**', '!**/vendor/**'],
+  collectCoverageFrom: ['**/*.{ts,}', '!**/node_modules/**', '!**/dist/**'],
   coverageThreshold: {
     global: {
       branches: 50,
