+++ tsconfig.json
@@ -10,13 +10,13 @@
     "target": "es5",
     "allowJs": true,
     "esModuleInterop": true,
-    "baseUrl": "src",
+    "baseUrl": ".",
     "paths": {
-      "@components/*": ["components/*"],
-      "@githubApi/*": ["github-Api/*"],
-      "@redux/*": ["redux/*"],
-      "@model/*": ["model/*"],
-      "@util/*": ["util/*"]
+      "@components/*": ["src/components/*"],
+      "@githubApi/*": ["src/github-Api/*"],
+      "@redux/*": ["src/redux/*"],
+      "@model/*": ["src/model/*"],
+      "@util/*": ["src/util/*"]
     }
   },
   "include": ["./src/**/*", "__test__/**/*"]
